# Conflicts:
#	gui/src/main/java/org/brabocoin/brabocoin/gui/BrabocoinGUI.java
#	gui/src/main/java/org/brabocoin/brabocoin/gui/view/MainView.java

+++ core/src/main/java/org/brabocoin/brabocoin/validation/block/rules/MaxSizeBlkRule.java
@@ -11,7 +11,7 @@ public class MaxSizeBlkRule extends BlockRule {
     @DescriptionField
     private int blockSize;
     @DescriptionField
-    private long consensusBlockSize;
+    private int consensusBlockSize;
 
     @Override
     public boolean isValid() {
